common: rename primitive_impl_t -> primitive_t
The idea is to use dnnl_primitive class as a user facing entity and
introduce an alias dnnl_primitive_iface_t for it to use it
internally, rather than primitive_t as it was before. The primitive_t
takes over responsibility of the primitive_impl_t and the primitive_impl_t
is deleted as it's no longer needed.
The same approach will be used for primitive_desc_t.

According to the existing namespace structure that approach looks reasonable as
introducing the `impl` suffix is redundant.

dnnl {
impl {
    struct primitive_t;
    strcut primitive_desc_t;
}
}

Rationale: according to the proposal on global primitive cache
there should be introduced primitive_desc_impl_t but following
the existing naming convention, introducing primitive_desc_impl_t
will cause massive renaming and make the names longer which is not
convenient as the names are used oftentimes.