benchdnn: move attr-scratchpad support inside drivers
Global settings causes issues when option is specified in a harness and
not reset by same input file back. However, resetting the option must be
done by running a problem with previously set scratchpad mode which is
an undesired limitation of the feature. Moving feature support under a
driver will reset it properly no matter what.