cpu: x64: gemm: use zero extend in copy_b_sum kernels
Matrix B is u8, so we need to use zero-extend instead of sign-extend.
This was fixed in the past for avx512, applying the fix for sse41, avx
and avx2.

For testing, having an element large than 127 in case of unsigned matrix
and a negative in case of signed matrix is enough catch this bug.