gpu: ocl: set f32 winograd to default to 6x3 method
This patch creates quite a bit of dead code, namely the logic for handling
non-fused Winograd algorithms, along with the 2x3 non-fused algorithm is no
longer use. This code is not removed at this time as Winograd is a work in
progress and some of this code may be needed again in the near future.