benchdnn: utils: compare: move post-ops check to diff
It happens that output from operation may be in certain range which may
be a range of values leading to catastrophic cancellation when specific
eltwise post-ops are used. To avoid such errors, we relax the condition
of such check making it diff < custom_trh. We rely our precise
validation of eltwise on eltwise driver where a wide range of inputs are
used and such ranges are validated separately.