benchdnn: move unimplemented status validation after pd creation
Before this change, both invalid cases and unimplemented cases were
validated upfront the pd creation which could lead to situations when
supported problems could be skipped. With new approach, invalid cases
are still checked before pd is created but unimplemented cases are
handled after pd was created. This allows to give a problem a chance to
be actually run if it becomes supported.