cpu: aarch64: add ACL PReLU primitive
This PR adds a PReLU primitive which makes use of Compute Library for the Arm® architecture (ACL), optimised for AArch64 targets. The datatype support is just for f32 and can only be used in forward mode.
Implementation follows a similar approach to #1281, including measures to avoid the parallelisation of small workloads. It builds upon the approach originally introduced in the PR #820 and RFC #795. The primitive offers a minimum of ~ x10 speedup
for large tensors (~1million f32s) and is comparable for very small tensors.

Co-authored-by: Louis Kaplan <louis.kaplan@arm.com>

- [X] Do all unit and benchdnn tests (`make test` and `make test_benchdnn_*`) pass locally for each commit?
- [X] Have you formatted the code using clang-format?

- [X] Have you submitted performance data that demonstrates performance improvements?