benchdnn: add INIT mode (`--mode=I`)
Add INIT mode for primitive initialization testing. Generally, with this
mode primitives are created but not executed and memory is described but
not allocated. A couple ways this mode may be useful:
 - check that expected implementations are dispatched by checking output
   implementation strings with verbose mode (level 5 or greater)
 - check that code generation during primitive creation is successful

In particular, x64 JIT kernels may fail during the code generation stage
if any immediate values exceed the range of signed 32-bit integers. This
can be hard to extensively validate as such cases tend to occur huge for
problems that hog resources. With INIT mode we do not have to run a full
primitive so testing becomes fast and easy.