x64: ip: use minimum os block under certain cases
Depending on the loop order and the os_block needed for parallelization,
it might be advantageous to use the register blocking on m-dimension as
os_block to give maximum parallelization + avoiding tail blocks at each
call of the kernel.

Small os_block implies many calls, therefore it would be better to avoid
kernel tails if possible.