cmake: aarch64: Defaults to a generic cpu target
This is equivalent to setting -mtune=generic -march=armv8-a.
This ensures no implementation specific tuning, or architectural
features beyond armv8-a are used,for portability across
AArch64 systems.The DNNL_ARCH_OPT_FLAGS build option can be used
to override these defaults to optimise for a specific cpu,
or revision of the Armv8 architecture.